+++ packages/component-library/src/components/MailchimpForm/MailchimpForm.tsx
@@ -156,7 +156,8 @@ const CustomForm = ({
             height: '100%',
             opacity: status === 'success' ? 1 : 0,
             pointerEvents: status === 'success' ? 'initial' : 'none'
-          }}>
+          }}
+        >
           {successMessage ? <Text body={successMessage} /> : <Box>Success</Box>}
         </Grid>
       </FormContainer>